bot/requirements.txt
aiogram==3.20.0.post0
httpx==0.28.1
python-dotenv==1.1.0